requirements.txt
pandas
Django
gunicorn
djangorestframework
drf-yasg
django-cors-headers
drf-yasg[validation]
flex
swagger_spec_validator
munch
psycopg2-binary
scipy
packaging
whitenoise
dj-database-url
# Deps para testes:
coverage
git+https://gitlab.com/pycqa/flake8#egg=flake8  # evita https://github.com/PyCQA/pycodestyle/issues/728
